move WordNetId
